refactor(server): move server package to internal
