+++ cmd/flipt/main.go
@@ -33,6 +33,10 @@ import (
 	"github.com/spf13/cobra"
 	"go.flipt.io/flipt/internal/config"
 	"go.flipt.io/flipt/internal/info"
+	"go.flipt.io/flipt/internal/server"
+	"go.flipt.io/flipt/internal/server/cache"
+	"go.flipt.io/flipt/internal/server/cache/memory"
+	"go.flipt.io/flipt/internal/server/cache/redis"
 	"go.flipt.io/flipt/internal/storage"
 	"go.flipt.io/flipt/internal/storage/sql"
 	"go.flipt.io/flipt/internal/storage/sql/mysql"
@@ -40,10 +44,6 @@ import (
 	"go.flipt.io/flipt/internal/storage/sql/sqlite"
 	"go.flipt.io/flipt/internal/telemetry"
 	pb "go.flipt.io/flipt/rpc/flipt"
-	"go.flipt.io/flipt/server"
-	"go.flipt.io/flipt/server/cache"
-	"go.flipt.io/flipt/server/cache/memory"
-	"go.flipt.io/flipt/server/cache/redis"
 	"go.flipt.io/flipt/swagger"
 	"go.flipt.io/flipt/ui"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
