+++ internal/server/cache/metrics.go
