+++ internal/server/rule_test.go
