+++ internal/server/segment.go
