+++ internal/server/segment_test.go
